+++ src/oms_sensemaking/geospatial/sensemakers/similar_tracks.py
@@ -94,7 +94,7 @@ class SimilarTracksSensemaker(Sensemaker):
         :param data: Track object to detect cotravels on
         :return: list[PotentialMatch] list of TopSimilar tracks
         """
-        LOGGER.debug(f"Looking for similar tracks to {data.node_id}")
+        LOGGER.debug("Looking for similar tracks to %s", data.node_id)
         # Update the config with specific geo settings
         self.config.update(config)
 
@@ -152,7 +152,7 @@ class SimilarTracksSensemaker(Sensemaker):
         union = len(ref_track_geohash_set.union(eval_track_geohash_set))
         score = intersection / union
         # can we do this
-        LOGGER.debug(f"Overall similarity for tracks {score}")
+        LOGGER.debug("Overall similarity for tracks %f", score)
         return ComparisonResult(track_uuid, score)
 
     @classmethod
@@ -311,13 +311,13 @@ class SimilarTracksSensemaker(Sensemaker):
             start_geojson = {"type": "Point", "coordinates": first}
 
             start_query = generate_query(Point.detection_time, start_geojson)
-            LOGGER.debug(f"Start bookend query {start_query.compile(db.bind, compile_kwargs={'literal_binds':True})}")
+            LOGGER.debug("Start bookend query %s", start_query.compile(db.bind, compile_kwargs={"literal_binds": True}))
             res = db.execute(start_query)
             start_groups = res.all()
 
             end_geojson = {"type": "Point", "coordinates": last}
             end_query = generate_query(desc(Point.detection_time), end_geojson)
-            LOGGER.debug(f"End bookend query {end_query.compile(db.bind, compile_kwargs={'literal_binds':True})}")
+            LOGGER.debug("End bookend query %s", end_query.compile(db.bind, compile_kwargs={"literal_binds": True}))
             res = db.execute(end_query)
             end_groups = res.all()